Register user using a server/webapp
GUI that allows the user to input their phone number and how often they want to be reminded to drink water. When pressing save the data is stored in a variable. Sending a post request to the server at /data puts the data stored in the variable into formData.json which can be read in the C program when a new RFID tag is detected.